# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "B2C Marketing Automation API",
    "description": "Complete REST API for the B2C Marketing Automation Platform. Manage contacts, workflows, deliveries, segments, audiences, custom objects, and more.",
    "version": "2.0.0",
    "contact": {
      "name": "Platform Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "List contacts",
        "description": "Get all contacts with pagination and filtering support.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated list of contacts" } }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create contact",
        "description": "Create a new B2C consumer profile.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContactInput" } } } },
        "responses": { "201": { "description": "Contact created" } }
      }
    },
...
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Contact details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update contact",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContactInput" } } } },
        "responses": { "200": { "description": "Contact updated" } }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contact",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Contact deleted" } }
      }
    },
...
    "/workflows": {
      "get": {
        "tags": ["Workflows"],
        "summary": "List workflows",
        "description": "Get all workflows with optional type/status filtering.",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["broadcast", "automated", "transactional"] } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "List of workflows" } }
      },
      "post": {
        "tags": ["Workflows"],
        "summary": "Create workflow",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkflowInput" } } } },
        "responses": { "201": { "description": "Workflow created" } }
      }
    },
...
    "/workflows/quick/send-now": {
      "post": {
        "tags": ["Workflows"],
        "summary": "Quick send now",
        "description": "Create and immediately activate a broadcast workflow.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Workflow created and activated" } }
      }
    },
    "/workflows/{id}": {
      "get": {
        "tags": ["Workflows"],
        "summary": "Get workflow",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow details" } }
      },
      "put": {
        "tags": ["Workflows"],
        "summary": "Update workflow",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow updated" } }
      },
      "delete": {
        "tags": ["Workflows"],
        "summary": "Delete workflow",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow deleted" } }
      }
    },
    "/workflows/{id}/activate": {
      "post": {
        "tags": ["Workflows"],
        "summary": "Activate workflow",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow activated" } }
      }
    },
    "/workflows/{id}/pause": {
      "post": {
        "tags": ["Workflows"],
        "summary": "Pause workflow",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow paused" } }
      }
    },
    "/workflows/{id}/complete": {
      "post": {
        "tags": ["Workflows"],
        "summary": "Complete workflow",
        "description": "Mark a broadcast workflow as complete.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow completed" } }
      }
    },
    "/workflows/{id}/archive": {
      "post": {
        "tags": ["Workflows"],
        "summary": "Archive workflow",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow archived" } }
      }
    },
    "/workflows/{id}/orchestration": {
      "get": {
        "tags": ["Workflows"],
        "summary": "Get workflow orchestration",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Orchestration canvas data" } }
      },
      "put": {
        "tags": ["Workflows"],
        "summary": "Save workflow orchestration",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Orchestration saved" } }
      }
    },
    "/workflows/{id}/report": {
      "get": {
        "tags": ["Workflows"],
        "summary": "Get workflow report",
        "description": "Get performance metrics and report data for a workflow.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Workflow report" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContactInput": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "country": { "type": "string" }
        },
        "required": ["email", "first_name", "last_name"]
      },
      "WorkflowInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["broadcast", "automated", "transactional"] },
          "channel": { "type": "string", "enum": ["email", "sms", "push"] },
          "subject": { "type": "string" },
          "content": { "type": "string" },
          "segment_id": { "type": "integer" },
          "audience_id": { "type": "integer" }
        },
        "required": ["name", "type"]
      },
...
    }
  }
}